requirements.txt
pytest-django==4.1.0
requests==2.25.0
pillow==8.0.1
sorl-thumbnail==12.7.0
django==3.0
djangorestframework==3.12.2
